Problème serveur

+++ controllers/Book.js
@@ -1,4 +1,5 @@
 const Book = require('../models/Book');
+const fs = require('fs');
 
 exports.createBook = (req, res, next) => {
     const bookObject = JSON.parse(req.body.book);
@@ -7,6 +8,8 @@ exports.createBook = (req, res, next) => {
     const book = new Book({
         ...bookObject,
         userId: req.auth.userId,
+        ratings: [],
+        averageRating: 0,
         imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
     });  
     thing.save()
@@ -15,24 +18,45 @@ exports.createBook = (req, res, next) => {
  };
 
 exports.modifyBook = (req, res, next) => {
-    Book.updateOne({ _id: req.params.id }, { ...req.body, id: req.params.id })
-        .then(() => res.status(200).json({message: 'Livre modifié !'}))
-        .catch(error => res.status(400).json({error}));
-};
+    const bookObject = req.file ? {
+        ...JSON.parse(req.body.book),
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}` 
+    } : { ...req.body };
 
-exports.ratingBook = (req, res, next) => {
-    // a faire //
+    delete bookObject._userId;
+    Book.findOne({_id: req.params.id})
+        .then((book) => {
+            if(book.userId != req.auth.userId) {
+                res.status(401).json({message: 'Non-autorisé !'})
+            } else {
+                Book.updateOne({_id: req.params.id}, {...bookObject, _id: req.params.id})
+                    .then(() => res.status(200).json({message: 'Objet modifié !'}))
+                    .catch(error => res.status(401).json({error}));
+            }
+        })
+        .catch(error => res.status(400).json({error}))
 };
 
 exports.deleteBook = (req, res, next) => {
-    Book.deleteOne({ _id: req.params.id })
-        .then(() => res.status(200).json({message: 'Livre supprimé !'}))
-        .catch(error => res.status(400).json({error}));
+    Book.findOne({ _id: req.params.id})
+        .then(book => {
+            if(book.userId != req.auth.userId) {
+                res.status(401).json({message: 'Non-autorisé !'})
+            } else {
+                const filename = book.imageUrl.split('/images/')[1];
+                fs.unlink(`images/${filename}`, () => {
+                    Book.deleteOne({_id: req.params.id})
+                        .then(() => res.status(200).json({message: 'Objet Supprimé !'}))
+                        .catch(error => res.status(401).json({error}));
+                });
+            }
+        })
+        .catch(error => res.status(500).json({error}));
 };
 
 exports.getOneBook = (req, res, next) => {
     Book.findOne({ _id: req.params.id })
-        .then(book => res.status(200).json(book))
+        .then((book) => res.status(200).json(book))
         .catch(error => res.status(404).json({error}));
 };
 
@@ -42,6 +66,11 @@ exports.getAllBooks = (req, res, next) => {
     .catch(error => res.status(400).json({error}));
 };
 
+exports.ratingBook = (req, res, next) => {
+    // a faire //
+};
+
 exports.getBestRatings = (req, res, next) => {
     // à faire //
 };
+
